improve design for view where you can't delete account yet
